% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{add_row}
\alias{add_row}
\title{Add a New Row}
\usage{
add_row(x, position = "first", session)
}
\arguments{
\item{x}{A data table-like object or a reactive expression returning a data table-like object.}

\item{position}{Character or numeric value indicating where the new row should be inserted. It can take one of the following values:
\itemize{
\item "first": Insert the new row at the beginning of the data table (top).
\item "last": Insert the new row at the end of the data table (bottom).
\item numeric: A numeric value representing the row number where the new row should be inserted. If the value is within the range of the current data table, the new row will be inserted at the specified position.
}}

\item{...}{Additional arguments not used in this function.}
}
\value{
A data table-like object with the new row added.
}
\description{
This function adds a new row to a dataframe. The new row is populated with default values based on the data types of the columns present in the input object.
}
\examples{
# Create a sample data table
dt <- data.table::data.table(
  x = c(1L, 2L, 3L),
  y = c(0.1, 0.2, 0.3),
  z = c("A", "B", "C"),
  flag = c(TRUE, FALSE, TRUE),
  date_val = as.Date(c("2023-01-01", "2023-02-01", "2023-03-01")),
  timestamp_val = as.POSIXct(c("2023-01-01 12:00:00", "2023-02-01 12:00:00", "2023-03-01 12:00:00"))
)

# Add a new row with default values at the beginning (top) of the data table
new_dt_first <- add_row(dt, position = "first")

# Add a new row with default values at the end (bottom) of the data table
new_dt_last <- add_row(dt, position = "last")

# Add a new row with default values at the specified position (between row 2 and row 3)
new_dt_custom <- add_row(dt, position = 2)

}
