% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apply-table-changes.R
\name{addCol}
\alias{addCol}
\title{Adds a column to a table.}
\usage{
addCol(table, ind, ct)
}
\arguments{
\item{table}{The htable data.frame}

\item{ind}{The 0 based index location to add the column}

\item{ct}{The number of columns to add}
}
\value{
The data.frame provided with the change applied.
}
\description{
Adds a column to a table.
}

