% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apply-table-changes.R
\name{addRow}
\alias{addRow}
\title{Adds a row to a table.}
\usage{
addRow(table, ind, ct)
}
\arguments{
\item{table}{The htable data.frame}

\item{ind}{The 0 based index location to add the row}

\item{ct}{The number of rows to add}
}
\value{
The data.frame provided with the change applied.
}
\description{
Adds a row to a table.
}

