% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-htable-delta.R
\name{compareNA}
\alias{compareNA}
\title{This function returns TRUE wherever elements are the same, including NA's.}
\usage{
compareNA(v1, v2)
}
\arguments{
\item{v1}{}

\item{v2}{}
}
\value{
boolean
}
\description{
This function returns TRUE wherever elements are the same, including NA's.
}
\seealso{
http://www.cookbook-r.com/Manipulating_data/Comparing_vectors_or_factors_with_NA/
}

