% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apply-table-changes.R
\name{delCol}
\alias{delCol}
\title{Deletes a column from a table.}
\usage{
delCol(table, ind, ct)
}
\arguments{
\item{table}{The htable data.frame}

\item{ind}{The 0 based index location to delete the column}

\item{ct}{The number of columns to delete}
}
\value{
The data.frame provided with the change applied.
}
\description{
Deletes a column from a table.
}

