% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apply-table-changes.R
\name{delRow}
\alias{delRow}
\title{Deletes a row from a table.}
\usage{
delRow(table, ind, ct)
}
\arguments{
\item{table}{The htable data.frame}

\item{ind}{The 0 based index location to delete the row}

\item{ct}{The number of rows to delete}
}
\value{
The data.frame provided with the change applied.
}
\description{
Deletes a row from a table.
}

