% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTable.R
\name{generate_table_body}
\alias{generate_table_body}
\title{generate the table body}
\usage{
generate_table_body(x, id_cols, col_types, skip_cols, table_id)
}
\arguments{
\item{x}{A data.frame or reactive object containing the data to be displayed in the table.}

\item{id_cols}{A numeric vector of column indices that should be displayed as static text.}

\item{skip_cols}{A numeric vector of column indices to skip during table generation.}

\item{table_id}{An optional ID for the table. If not provided, the ID will default to the name of the input data.}
}
\description{
generate the table body
}
