% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTable.R
\name{get_column_input_type}
\alias{get_column_input_type}
\title{Get the input type for a given column class.}
\usage{
get_column_input_type(column_class)
}
\arguments{
\item{column_class}{A character string specifying the class of the column.}
}
\value{
A character string representing the input type for the given column class.
}
\description{
This function maps a given column class to an appropriate input type for HTML forms.
The input type is selected based on the column class and is returned as a character string.
If the column class is "character" or "factor", the input type will be "text".
If the column class is "logical", the input type will be "checkbox".
If the column class is "integer" or "numeric", the input type will be "number".
If the column class is "Date", the input type will be "date".
If the column class is "POSIXct" or "POSIXt", the input type will be "datetime-local".
If the column class is "times", the input type will be "time".
For any other column class, the default input type will be "text".
}
\examples{
# Example 1
get_column_input_type("character")
# Output: "text"

# Example 2
get_column_input_type("numeric")
# Output: "number"

}
