% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{hideRowsR}
\alias{hideRowsR}
\title{Hide table rows in a Shiny App}
\usage{
hideRowsR(table_id, hideIndex, session = getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{A string specifying the ID of the table that needs to be sorted. If used in a module you may need to add the namespace.}

\item{session}{A reactive domain (usually from \code{shiny::getDefaultReactiveDomain()}), representing the Shiny session. Defaults to the current session.}
}
\description{
This function sends a custom message to the Shiny session to hide rows from a table
This function is useful because the hidden rows are unaffected by search, and also
allow you to "filter" a table without interrupting the i,j index.
}
\details{
It is important to note that the rows are hidden, not removed, and in any
case they are rendered before being hidden. Any information
contained in rows is still available as html, and so is not secure.
}
