% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_scripts_documentation.R
\docType{data}
\name{inputChange}
\alias{inputChange}
\title{inputChange js function}
\format{
An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
inputChange
}
\description{
inputChange js function
}
\keyword{datasets}
