% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaged_javascript.R
\docType{data}
\name{js_handle_input_change}
\alias{js_handle_input_change}
\title{Handle input change javascript}
\format{
An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
js_handle_input_change
}
\description{
This is the entire tags$script(HTML(...)) object. This object is inserted after
the table is generated.
}
\keyword{datasets}
