% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaged_javascript.R
\docType{data}
\name{js_helpers}
\alias{js_helpers}
\title{js helpers}
\format{
An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
js_helpers
}
\description{
js helpers
}
\keyword{datasets}
