% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{remove_row}
\alias{remove_row}
\title{Delete a Row}
\usage{
remove_row(x, position)
}
\arguments{
\item{x}{A data table-like object or a reactive expression returning a data table-like object.}

\item{position}{Numeric value representing the row number to be deleted.}
}
\value{
A data table-like object with the specified row removed.
}
\description{
This function removes a row from a dataframe object based on the specified row number.
}
\examples{
# Create a sample data table
dt <- data.table::data.table(
  x = c(1L, 2L, 3L),
  y = c(0.1, 0.2, 0.3),
  z = c("A", "B", "C"),
  flag = c(TRUE, FALSE, TRUE),
  date_val = as.Date(c("2023-01-01", "2023-02-01", "2023-03-01")),
  timestamp_val = as.POSIXct(c("2023-01-01 12:00:00", "2023-02-01 12:00:00", "2023-03-01 12:00:00"))
)

# Delete row 2 from the data table
new_dt <- delete_row(dt, row_num = 2)

}
