% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_output.R
\name{rendershinyTable}
\alias{rendershinyTable}
\title{Custom Render Function for Shiny Tables}
\usage{
rendershinyTable(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns the table to be rendered.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This is useful if you want to save an expression in a variable.}
}
\value{
A function that returns a list containing the HTML of the table and any additional input scripts.
}
\description{
Render a shiny table using a custom output element.
}
