% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R, R/shiny_module.R,
%   R/shinytest.R
\name{run_test}
\alias{run_test}
\title{Run a test of the shiny table}
\usage{
run_test(mode = "data.frame")

run_test(mode = "data.frame")

run_test(mode = "data.frame")
}
\description{
Run a test of the shiny table

Run a test of the shiny table

Run a test of the shiny table
}
