% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{searchTableR}
\alias{searchTableR}
\title{Search Table in a Shiny App}
\usage{
searchTableR(tableId, searchString, session = getDefaultReactiveDomain())
}
\arguments{
\item{tableId}{A string specifying the ID of the table that needs to be sorted. If used in a module you may need to add the namespace.}

\item{searchString}{A string representing the search query that will be used to filter the table.}

\item{session}{A reactive domain (usually from \code{shiny::getDefaultReactiveDomain()}), representing the Shiny session. Defaults to the current session.}
}
\value{
Invisible NULL.
}
\description{
This function sends a custom message to the Shiny session to perform a search within a table based on the provided search string.
}
\examples{
\dontrun{
  searchTableR("myTableID", "searchQuery")
}

}
\seealso{
\code{\link[shiny]{session}}
\code{\link[shiny]{getDefaultReactiveDomain}}
}
