% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-htable-types.R
\name{setHtableClass}
\alias{setHtableClass}
\title{Covert htable output matrix to data.frame using classes of model data.frame}
\usage{
setHtableClass(data, old)
}
\arguments{
\item{data}{htable matrix}

\item{old}{original data.frame}
}
\value{
data.frame
}
\description{
Covert htable output matrix to data.frame using classes of model data.frame
}
\seealso{
https://stackoverflow.com/questions/9214819/supply-a-vector-to-classes-of-dataframe
}

