% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{shiny_table}
\alias{shiny_table}
\title{Create an editable HTML table}
\usage{
shiny_table(
  x,
  table_id = NULL,
  type_list = NULL,
  id_cols = 1,
  skip_cols = NULL,
  id = NULL,
  ...
)
}
\arguments{
\item{x}{a data.frame or reactive object}

\item{table_id}{the id of the table, defaults to x}

\item{type_list}{list of column input types of the format \code{list(text = 2, checkbox = c(3, 5))}. Columns are guessed by \code{shinyTable:::get_column_input_type}. Arguments override guesses}

\item{...}{}
}
\description{
Create an editable HTML table
}
