% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyTable-package.R
\docType{package}
\name{shinyTable-package}
\alias{shinyTable}
\alias{shinyTable-package}
\title{Handsontable for Shiny}
\description{
Handsontable wrappers for Shiny that make reading and writing
tabular data easy from Shiny.
}

