% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTable.R
\name{shinyTable}
\alias{shinyTable}
\title{Create an editable HTML table}
\usage{
shinyTable(
  x,
  table_id = NULL,
  id_cols = 1,
  type_list = NULL,
  skip_cols = NULL,
  id = NULL,
  ...
)
}
\arguments{
\item{x}{a data.frame or reactive object}

\item{table_id}{the id of the table, defaults to x}

\item{id_cols}{a numeric vector of columns that are displayed as static text}

\item{type_list}{list of column input types of the format \code{list(text = 2, checkbox = c(3, 5))}. Columns are guessed by \code{shinyTable:::get_column_input_type}. Arguments override guesses}

\item{skip_cols}{a numeric vector of columns to skip}

\item{id}{if used in a shiny module, the module id}

\item{...}{not used yet but don't knock it}
}
\description{
Create an editable HTML table
}
