% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_output.R
\name{shinyTableOutput}
\alias{shinyTableOutput}
\title{Custom Output for Shiny Tables}
\usage{
shinyTableOutput(id)
}
\arguments{
\item{id}{A character string indicating the ID of the output div.}
}
\value{
A Shiny tagList containing the custom table div and JavaScript dependencies.
}
\description{
Generate a custom shiny table output element.
}
