% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shinyTableServer}
\alias{shinyTableServer}
\title{Server logic for a shinyTable instance}
\usage{
shinyTableServer(
  id,
  x,
  table_id = NULL,
  id_cols = NULL,
  type_list = NULL,
  col_names = NULL,
  skip_cols = NULL,
  sortable = NULL,
  searchable = FALSE,
  uid_cols = NULL,
  mode = "inputs",
  ...
)
}
\arguments{
\item{id}{The module ID.}

\item{x}{A data frame or reactive data frame containing the data for the table.}

\item{table_id}{An optional ID for the table. If not provided, a default ID will be used.}

\item{id_cols}{A numeric vector of column indices to be displayed as static text.}

\item{type_list}{A list specifying input types for specific columns.}

\item{col_names}{A character vector specifying custom column names for the table headers.}

\item{skip_cols}{A numeric vector of column indices to skip during table generation.}

\item{sortable}{A flag indicating if the table should be sortable (TRUE or FALSE).}

\item{searchable}{A flag indicating if the table should be searchable (TRUE or FALSE).}

\item{uid_cols}{A numeric vector of column indices to be used as unique identifiers for each row.}

\item{mode}{The mode for the table, either "data.frame", "inputs", or "both".}

\item{...}{Additional arguments (currently not used).}
}
\value{
The module server function that defines the behavior of the shinyTable instance.
}
\description{
This module provides the server logic for a shinyTable instance, which allows you to create and manage an editable HTML table within a Shiny application.
}
\examples{

# Use the module in the UI
ui <- fluidPage(
  shinyTableUI("table_module"),
  # Other UI elements...
)

# Define server logic for the UI
server <- function(input, output, session) {
  shinyTableServer("table_module", data.frame(Name = c("Alice", "Bob"), Age = c(25, 30)), mode = "inputs")
  # Other server logic...
}

shinyApp(ui, server)

}
