% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R, R/shiny_module.R,
%   R/shinytest.R
\name{shinyTableUI}
\alias{shinyTableUI}
\title{UI for an interactive shinyTable module}
\usage{
shinyTableUI(id, verbose = interactive())

shinyTableUI(id, verbose = interactive())

shinyTableUI(id, verbose = interactive())
}
\arguments{
\item{id}{module ID}
}
\description{
UI for an interactive shinyTable module

UI for an interactive shinyTable module

UI for an interactive shinyTable module
}
