% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{sortTableR}
\alias{sortTableR}
\title{Sort Table in a Shiny App}
\usage{
sortTableR(table_id, sortDict, session = getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{A string specifying the ID of the table that needs to be sorted. If used in a module you may need to add the namespace.}

\item{sortDict}{A list specifying the sorting parameters. The names of the list are the column names and the list values are the sorting directions (e.g., 'asc' for ascending).}

\item{session}{A reactive domain (usually from \code{shiny::getDefaultReactiveDomain()}), representing the Shiny session. Defaults to the current session.}
}
\value{
Invisible NULL.
}
\description{
This function sends a custom message to the Shiny session to sort a table based on the given parameters.
}
\examples{
\dontrun{
  sortTableR("myTableID", list(column1 = 'asc', column2 = 'desc'))
}

}
\seealso{
\code{\link[shiny]{session}}
\code{\link[shiny]{getDefaultReactiveDomain}}
}
