% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{table_html}
\alias{table_html}
\title{Title}
\usage{
table_html(x, table_id = NULL, type_list, ...)
}
\arguments{
\item{x}{a data.frame or reactive object}

\item{table_id}{the id of the table, defaults to x}

\item{type_list}{list of column input types of the format \code{list(text = 2, checkbox = c(3, 5))}. Columns are guessed by \code{shinyTable:::get_column_input_type}. Arguments to type_list override guesses}

\item{...}{}
}
\description{
Title
}
