% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-htable-types.R
\name{toCls}
\alias{toCls}
\title{Convert to specified class}
\usage{
toCls(x, cls)
}
\arguments{
\item{x}{vector}

\item{cls}{character}
}
\value{
converted vector
}
\description{
Convert to specified class
}
\seealso{
https://stackoverflow.com/questions/9214819/supply-a-vector-to-classes-of-dataframe
}

