% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{uid}
\alias{uid}
\title{Create a uid based on system time}
\usage{
uid(...)
}
\arguments{
\item{...}{additional uids to avoid time confrontations. For example, a user
uid in addition to the system time to ensure that simultaneous edits by two
users are still unique records

Note that the hash is created by crc32, and is not recommended to be used to
hide sensitive information like an email or other identifying records}
}
\description{
Create a uid based on system time
}
