% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_scripts_documentation.R
\docType{data}
\name{updateInput}
\alias{updateInput}
\title{updateInput js function}
\format{
An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
updateInput
}
\description{
updateInput js function
}
\keyword{datasets}
