% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{updateInputR}
\alias{updateInputR}
\title{update a table's input and fire off table input with action: value_change}
\usage{
updateInputR(tableId, i, j, value)
}
\arguments{
\item{tableId}{A string specifying the ID of the table that needs to be sorted. If used in a module you may need to add the namespace.}

\item{i}{the i index / attribute}

\item{j}{the j index / attribute}

\item{value}{the new value to replace the input}
}
\description{
update a table's input and fire off table input with action: value_change
}
