% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-htable-types.R
\name{getHtableTypes}
\alias{getHtableTypes}
\title{Calculate HTable-Compatible Type Names}
\usage{
getHtableTypes(data)
}
\arguments{
\item{data}{The data.frame to analyze.}
}
\description{
Given a data.frame, extract the classes of each column and convert them to
the nomenclature used by Handsontable.
}
\author{
Jeff Allen \email{jeff@trestletech.com}
}

