\name{htable}
\alias{htable}
\title{Render a Handsontable}
\usage{
  htable(outputId)
}
\arguments{
  \item{outputId}{The ID of the \code{glOutput} associated
  with this element}
}
\description{
  Render a Handsontable in an application page.
}
\author{
  Jeff Allen \email{jeff@trestletech.com}
}

