% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_scripts_documentation.R
\docType{data}
\name{searchTable}
\alias{searchTable}
\title{searchTable js function}
\format{
An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
searchTable
}
\description{
searchTable js function
}
\keyword{datasets}
