% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTable-package.R
\docType{package}
\name{shinyTable-package}
\alias{shinyTable}
\alias{shinyTable-package}
\title{shinyTable: Simple html tables built for shiny + CRUD operations}
\description{
Generate html tables uing html inputs and an [i,j] grid for a more R native approach to interactive table editing in R's shiny framework.
}
\author{
\strong{Maintainer}: Kent Orr \email{orrkent@pawpawanalytics.com}

}
\keyword{internal}
