% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTable.R
\name{shinyTable}
\alias{shinyTable}
\title{Create an editable HTML table}
\usage{
shinyTable(
  x,
  table_id = NULL,
  id_cols = 1,
  type_list = NULL,
  col_names = NULL,
  skip_cols = NULL,
  sortable = "asc",
  searchable = TRUE,
  ns = NULL,
  ...
)
}
\arguments{
\item{x}{A data.frame or reactive object containing the data to be displayed in the table.}

\item{table_id}{An optional ID for the table. If not provided, the ID will default to the name of the input data.}

\item{id_cols}{A numeric vector of column indices that should be displayed as static text.}

\item{type_list}{A list specifying input types for specific columns. The format should be \code{list(input_type = c(column_indices))}. Column input types are guessed using \code{shinyTable:::get_column_input_type}, and this argument can be used to override the guesses. To set step use 'number-\if{html}{\out{<step>}}' as in 'number-.01'}

\item{col_names}{A character vector specifying custom column names for the table headers \code{new_name = old_name}. If not provided, column names from the input data will be used.}

\item{skip_cols}{A numeric vector of column indices to skip during table generation.}

\item{sortable}{either "asc" or "desc" or FALSE, giving sort order or no sort at all}

\item{searchable}{A boolean to indicate if a search box should be implemented.}

\item{ns}{The namespace of the Shiny module if used within a module context.}

\item{...}{Additional arguments (currently not used).}
}
\value{
An HTML table with interactive input cells and static text cells based on the provided data and parameters.
}
\description{
This function generates an editable HTML table based on the provided data. It can be used in Shiny applications to display and interact with tabular data.
}
\examples{
# Generate a simple editable table with default settings
shinyTable(data.frame(A = 1:5, B = 6:10))

# Generate a table with custom column names and input types
shinyTable(data.frame(Name = c("Alice", "Bob"), Age = c(25, 30)), col_names = c("Person", "Years"), type_list = list(text = 1, numeric = 2))

# Generate a table with specified static text columns and skipped columns
shinyTable(data.frame(ID = 1:3, Name = c("Alice", "Bob", "Carol"), Value = c(10, 20, 30)), id_cols = 1, skip_cols = 3)

}
