% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shinyTableServer}
\alias{shinyTableServer}
\title{Server logic for a shinyTable instance}
\usage{
shinyTableServer(
  id,
  x,
  mode = "inputs",
  table_id = NULL,
  id_cols = 1,
  skip_cols = NULL,
  type_list = NULL,
  ...
)
}
\arguments{
\item{id}{module ID}

\item{x}{a dataframe or reactive dataframe}

\item{mode}{either "data.frame" or "inputs"}
}
\description{
Server logic for a shinyTable instance
}
