% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shinyTableUI}
\alias{shinyTableUI}
\title{Shiny Table UI}
\usage{
shinyTableUI(
  id,
  shiny_search = TRUE,
  shiny_sort = TRUE,
  add_remove = c("Add Row", "Remove Row")
)
}
\arguments{
\item{id}{A string specifying the ID for the UI element.}

\item{shiny_search}{A logical value indicating whether the search functionality should be enabled via shiny. Defaults to TRUE.}

\item{shiny_sort}{A logical value indicating whether the sorting functionality should be enabled via shiny. Defaults to TRUE.}

\item{add_remove}{A character vector with two elements: "Add Row" and "Remove Row". Indicates the labels for the add and remove buttons. Can be NULL to disable the add/remove feature. Defaults to c("Add Row", "Remove Row").}
}
\value{
A tag list that can be added to a Shiny UI.
}
\description{
This function creates a user interface for a Shiny table with optional features like search, sort, and add/remove rows being implemented in Shiny.
}
\examples{
\dontrun{
  ui <- fluidPage(
    shinyTableUI("tableID", shiny_search = TRUE, shiny_sort = TRUE, add_remove = c("Add Row", "Remove Row"))
  )
}

}
\seealso{
\code{\link[shiny]{tagList}}
\code{\link[shiny]{textInput}}
\code{\link[shiny]{uiOutput}}
\code{\link[shiny]{actionButton}}
}
