% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shinyTableUI}
\alias{shinyTableUI}
\title{UI for an interactive shinyTable module}
\usage{
shinyTableUI(
  id,
  add_remove = TRUE,
  verbose = interactive(),
  add_text = "Add Row",
  remove_text = "Remove Row"
)
}
\arguments{
\item{id}{module ID}

\item{add_remove}{should there be add/remove row buttons}

\item{verbose}{print console?}

\item{add_text}{button label for add row}

\item{remove_text}{button label for remove row}
}
\description{
UI for an interactive shinyTable module
}
