% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaged_javascript.R
\docType{data}
\name{sortTable}
\alias{sortTable}
\title{sortTable}
\format{
An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
sortTable
}
\description{
sortTable
}
\keyword{datasets}
